# api-docs.json
{"openapi":"3.0.1","info":{"title":"Music Shop API","description":"An API to run a music shop","version":"1.0.0"},"servers":[{"url":"http://localhost:8080"}],"tags":[{"name":"api","description":"main resource 'product(s)'"},{"name":"user","description":"Available to registered and authenticated users"},{"name":"admin","description":"Available only to administrator"},{"name":"registration","description":"Publicly available endpoints for user registration"}],"paths":{"/api/products/{id}":{"get":{"tags":["api","user"],"summary":"Gets a single product","description":"Retrieves product from DB by provided id","operationId":"getProductById","parameters":[{"name":"id","in":"path","description":"Valid id number","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"responses":{"200":{"description":"Successful response","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}},"400":{"description":"Not found exception","content":{"*/*":{"example":"\"status\": 400,\n\"message\": \"You must use valid category!\",\n\"path\": \"/api/products\",\n\"timeStamp\": \"2022-08-05 11:27:40\""}}}}},"put":{"tags":["api","admin"],"summary":"Updates a single product","description":"Updates a product selected by provided id. Input is defined by a DTO and it is validated against set of constraints","operationId":"updateProduct","parameters":[{"name":"id","in":"path","description":"Valid id number","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}},"delete":{"tags":["api","admin"],"summary":"Deletes a single product from DB","description":"Deletes a single product by provided id","operationId":"deleteProductById","parameters":[{"name":"id","in":"path","description":"Valid id number","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"responses":{"200":{"description":"OK"}}}},"/api/product-categories/{id}":{"put":{"tags":["api","admin"],"summary":"Updates a product category","description":"Updates a product category. Input is validated against a number of constraints","operationId":"updateCategory","parameters":[{"name":"id","in":"path","description":"Valid id number","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCategory"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductCategory"}}}}}},"delete":{"tags":["api","admin"],"summary":"Deletes a product category","description":"Removes a product category from DB","operationId":"deleteProductCategory","parameters":[{"name":"id","in":"path","description":"Valid id number","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"responses":{"200":{"description":"OK"}}}},"/api/artists/{id}":{"put":{"tags":["api","admin"],"summary":"Updates an artist","description":"Updates an artist. Input is validated against a number of constraints","operationId":"updateArtist","parameters":[{"name":"id","in":"path","description":"Valid id number","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Artist"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Artist"}}}}}},"delete":{"tags":["api","admin"],"summary":"Deletes an artist","description":"Removes an artist from DB","operationId":"deleteArtist","parameters":[{"name":"id","in":"path","description":"Valid id number","required":true,"schema":{"type":"integer","format":"int64"},"example":3}],"responses":{"200":{"description":"OK"}}}},"/api/registration":{"post":{"tags":["registration","user"],"summary":"Registers a new user","description":"Registers a new user. Input is validated against a number of constraints","operationId":"registerUser","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/registration/savePassword":{"post":{"tags":["registration","user"],"summary":"Saves the password","description":"Saves the password. Expects token in url. Input is validated against a set of constraints","operationId":"savePassword","parameters":[{"name":"token","in":"query","description":"Valid token","required":true,"schema":{"type":"string"},"example":"0319c819-6ca3-4391-8807-91c4d0157f0b"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordSaveDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/registration/resetPassword":{"post":{"tags":["registration","user"],"summary":"Resets a password","description":"Resets a password. A link with a token is sent to user's email","operationId":"resetPassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PasswordChangeByEmailDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/registration/changePassword":{"post":{"tags":["registration","user"],"summary":"Changes the password","description":"Changes the password. Input is validated against a set of constraints","operationId":"changePassword","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChangePasswordDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/products":{"get":{"tags":["api","user"],"summary":"Gets all the products","description":"Retrieves product from DB using DTO and pages","operationId":"getProducts","parameters":[{"name":"page","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductDTO"}}}}}},"post":{"tags":["api","admin"],"summary":"Saves a single product","description":"Saves a product to DB. Input is defined by a DTO and it is validated against set of constraints","operationId":"saveProduct","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}}},"/api/product-categories":{"get":{"tags":["api","admin"],"summary":"Gets a list of product categories","description":"Retrieves a list of product categories","operationId":"getCategories","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProductCategory"}}}}}}},"post":{"tags":["api","admin"],"summary":"Saves a new product category","description":"Saves a new product category. Input is validated against a number of constraints","operationId":"saveProductCategory","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProductCategory"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductCategory"}}}}}}},"/api/artists":{"get":{"tags":["api","admin"],"summary":"Gets a list of artist","description":"Retrieves a list of artist from DB","operationId":"getAllArtists","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Artist"}}}}}}},"post":{"tags":["api","admin"],"summary":"Saves a new artist","description":"Saves a new artist. Input is validated against a number of constraints","operationId":"saveArtist","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Artist"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Artist"}}}}}}},"/api/registration/verifyRegistration":{"get":{"tags":["registration","user"],"summary":"Verifies a registration","description":"Verifies a registration. Confirmation link with a token is sent to user's email","operationId":"verifyRegistration","parameters":[{"name":"token","in":"query","description":"Valid token","required":true,"schema":{"type":"string"},"example":"0319c819-6ca3-4391-8807-91c4d0157f0b"}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/registration/resendToken":{"get":{"tags":["registration","user"],"summary":"Resends the token","description":"Resends the token in case it has expired","operationId":"resendVerificationToken","parameters":[{"name":"token","in":"query","description":"Valid token","required":true,"schema":{"type":"string"},"example":"0319c819-6ca3-4391-8807-91c4d0157f0b"}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/products/search/{query}":{"get":{"tags":["api","user"],"summary":"Gets a page of products by a query","description":"Retrieves a page of products by a query. Searches through product name field.Input query is case insensitive.","operationId":"findByNameContainingIgnoreCase","parameters":[{"name":"query","in":"path","description":"String of characters","required":true,"schema":{"type":"string"},"example":"met"},{"name":"page","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductDTO"}}}}}}},"/api/products/range":{"get":{"tags":["api","user"],"summary":"Gets a page of products between a given price range","description":"Retrieves a page of products between a given price range. Params have default values so that this endpoint returns the page in any case.","operationId":"getRangedPricedProducts","parameters":[{"name":"start","in":"query","description":"Starting price range, inclusive","required":false,"schema":{"type":"number","default":0},"example":10.0},{"name":"end","in":"query","description":"Ending price range, inclusive","required":false,"schema":{"type":"number","default":10000},"example":30.0},{"name":"page","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductDTO"}}}}}}},"/api/products/most-expensive":{"get":{"tags":["api","user"],"summary":"Gets a single most expansive product in DB","description":"Retrieves a single most expansive product","operationId":"getMostExpensiveProductInTheCatalogue","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProductDTO"}}}}}}},"/api/products/category/name/{categoryName}":{"get":{"tags":["api","user"],"summary":"Gets a page of products by a category","description":"Retrieves a page of products by a category that is identified by provided category name","operationId":"findByCategoryName","parameters":[{"name":"categoryName","in":"path","description":"Valid category name","required":true,"schema":{"type":"string"},"example":"CD"},{"name":"page","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductDTO"}}}}}}},"/api/products/category/id/{categoryId}":{"get":{"tags":["api","user"],"summary":"Gets a page of products by a category","description":"Retrieves a page of products by a category that is identified by provided category id","operationId":"findByCategoryId","parameters":[{"name":"categoryId","in":"path","description":"Valid categoryId number","required":true,"schema":{"type":"integer","format":"int64"},"example":3},{"name":"page","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductDTO"}}}}}}},"/api/products/artist/{name}":{"get":{"tags":["api","user"],"summary":"Gets a page of products by an atist","description":"Retrieves a a page of products by an artist name","operationId":"findByArtist","parameters":[{"name":"name","in":"path","description":"Valid artist name","required":true,"schema":{"type":"string"},"example":"Metallica"},{"name":"page","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageProductDTO"}}}}}}}},"components":{"schemas":{"ProductDTO":{"required":["active","artist","category","imageUrl","name","unitPrice","unitsInStock"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"sku":{"type":"string"},"name":{"maxLength":50,"minLength":1,"type":"string"},"description":{"maxLength":350,"minLength":0,"type":"string"},"imageUrl":{"maxLength":50,"minLength":3,"type":"string"},"unitPrice":{"maximum":1000,"minimum":1,"type":"number"},"active":{"type":"boolean"},"unitsInStock":{"maximum":100,"minimum":0,"type":"integer","format":"int32"},"artist":{"maxLength":50,"minLength":1,"type":"string"},"category":{"maxLength":4,"minLength":2,"type":"string"}}},"ProductCategory":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":4,"minLength":2,"type":"string"}}},"Artist":{"required":["name"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"maxLength":50,"minLength":1,"type":"string"}}},"UserDTO":{"required":["email","firstName","lastName","password"],"type":"object","properties":{"firstName":{"maxLength":25,"minLength":2,"type":"string"},"lastName":{"maxLength":25,"minLength":2,"type":"string"},"email":{"type":"string"},"password":{"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,20}$","type":"string"},"matchingPassword":{"type":"string"},"passwordsEqual":{"type":"boolean"}}},"PasswordSaveDTO":{"required":["newPassword"],"type":"object","properties":{"newPassword":{"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,20}$","type":"string"}}},"PasswordChangeByEmailDTO":{"required":["email"],"type":"object","properties":{"email":{"type":"string"}}},"ChangePasswordDTO":{"required":["email","newPassword","oldPassword"],"type":"object","properties":{"email":{"type":"string"},"oldPassword":{"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,20}$","type":"string"},"newPassword":{"pattern":"^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,20}$","type":"string"},"matchingPassword":{"type":"string"},"passwordsEqual":{"type":"boolean"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageProductDTO":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/ProductDTO"}},"number":{"type":"integer","format":"int32"},"sort":{"$ref":"#/components/schemas/Sort"},"first":{"type":"boolean"},"last":{"type":"boolean"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"offset":{"type":"integer","format":"int64"},"sort":{"$ref":"#/components/schemas/Sort"},"pageSize":{"type":"integer","format":"int32"},"pageNumber":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"}}},"Sort":{"type":"object","properties":{"empty":{"type":"boolean"},"sorted":{"type":"boolean"},"unsorted":{"type":"boolean"}}}},"securitySchemes":{"BearerJWT":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}